move and rename toggleMouseHandlers
